+++ github/Issue.py
@@ -18,7 +18,10 @@
 # Copyright 2016 Peter Buckley <dx-pbuckley@users.noreply.github.com>          #
 # Copyright 2017 Nicolas Agustín Torres <nicolastrres@gmail.com>              #
 # Copyright 2017 Simon <spam@esemi.ru>                                         #
+# Copyright 2018 Shinichi TAMURA <shnch.tmr@gmail.com>                         #
+# Copyright 2018 Steve Kowalik <steven@wedontsleep.org>                        #
 # Copyright 2018 Wan Liuyang <tsfdye@gmail.com>                                #
+# Copyright 2018 per1234 <accounts@perglass.com>                               #
 # Copyright 2018 sfdye <tsfdye@gmail.com>                                      #
 #                                                                              #
 # This file is part of PyGithub.                                               #
@@ -52,6 +55,8 @@ import github.Milestone
 import github.IssueComment
 import github.IssuePullRequest
 import github.Reaction
+
+import Consts
 
 
 class Issue(github.GithubObject.CompletableGithubObject):
@@ -392,7 +397,8 @@ class Issue(github.GithubObject.CompletableGithubObject):
             github.IssueEvent.IssueEvent,
             self._requester,
             self.url + "/events",
-            None
+            None,
+            headers={'Accept': Consts.mediaTypeLockReasonPreview}
         )
 
     def get_labels(self):
@@ -462,7 +468,7 @@ class Issue(github.GithubObject.CompletableGithubObject):
             self._requester,
             self.url + "/reactions",
             None,
-            headers={'Accept': 'application/vnd.github.squirrel-girl-preview'}
+            headers={'Accept': Consts.mediaTypeReactionsPreview}
         )
 
     def create_reaction(self, reaction_type):
@@ -482,7 +488,7 @@ class Issue(github.GithubObject.CompletableGithubObject):
             "POST",
             self.url + "/reactions",
             input=post_parameters,
-            headers={'Accept': 'application/vnd.github.squirrel-girl-preview'}
+            headers={'Accept': Consts.mediaTypeReactionsPreview}
         )
         return github.Reaction.Reaction(self._requester, headers, data, completed=True)